Tried to simplify install

+++ install.sh
@@ -1,11 +1,19 @@
 #!/usr/bin/env bash
 
+if [ ! -f $HOME/.ssh/id_rsa.pub ]; then
+
+    echo "Error: No RSA SSH key found at $HOME/.ssh/id_rsa.pub";
+    echo "Create a key with 'ssh-keygen -t rsa' before running install";
+    exit 0;
+
+fi
+
 if [ ! -d aicex ]; then
 
    git clone https://github.com/wulffern/aicex.git
    cd aicex
    mkdir .ssh
-   cat ~/.ssh/id_rsa.pub > .ssh/authorized_keys
+   cat $HOME/.ssh/id_rsa.pub > .ssh/authorized_keys
    docker run --rm  -p 2022:22 -v `pwd`:/home/aicex/ -i wulffern/aicex:latest &
 
 else
